refactor: コンポーネント名の変更
